refactor(control): rename tools ax to tools control, commander breakdown
src/ax becomes src/control with commands/ split: discovery (list/tree/
find/window/attrs/actions/preflight), interact (get/set/press/perform/
focus/click/type/hotkey/screenshot), state (snapshot/restore/build),
run (plan runner). Shared targeting options/args/label dedupe into
lib/target.ts. The capture runner moves to lib/capture-with-actions.ts,
exposed verbatim as 'tools control capture <args>' via raw-argv
passthrough (own argv parsing, flags untouched). tools macos control
replaces tools macos ax. Behavior frozen; capture runner brought under
repo lint (biome + SafeJSON) and tsconfig (two latent type errors:
missing onError on marker variants, unguarded axId in osascript
fallback).
